refactor: Hook 개선
