+++ src/hooks/useImageUpload.ts
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import uploadImage from '../common/api/imageApi';
+import { useMutation } from '@tanstack/react-query';
 
 /**
  * useImageUpload 훅
@@ -13,14 +14,14 @@ import uploadImage from '../common/api/imageApi';
  * - uploadSelectedImage: 서버에 이미지 업로드하고 imgId 반환
  */
 
+/**
+ * 파일 선택 시 미리보기 처리
+ * @param file 선택된 파일
+ */
 const useImageUpload = () => {
   const [imageFile, setImageFile] = useState<File | null>(null);
   const [imageUrl, setImageUrl] = useState<string>('');
 
-  /**
-   * 파일 선택 시 미리보기 처리
-   * @param file 선택된 파일
-   */
   const handleImageChange = (file: File | null) => {
     setImageFile(file);
     if (file) {
@@ -32,27 +33,33 @@ const useImageUpload = () => {
   };
 
   /**
-   * 선택된 이미지를 서버에 업로드하고 imgId 반환
-   * @returns number | null
+   * 이미지 파일을 서버 업로드 후 imgId 반환을 관리하는 커스텀 훅
+   */
+  const mutation = useMutation({
+    mutationFn: async (imageFile: File) => {
+      const imgIdStr = await uploadImage(imageFile);
+      const imgId = Number(imgIdStr);
+      if (isNaN(imgId)) {
+        throw new Error('업로드된 이미지 ID가 유효하지 않습니다.');
+      }
+      return imgId;
+    },
+  });
+
+  /**
+   * 선택된 이미지 업로드 실행 호출
    */
   const uploadSelectedImage = async (): Promise<number | null> => {
     if (!imageFile) return null;
-
-    const uploadResult = await uploadImage(imageFile);
-    const imgId = Number(uploadResult);
-
-    if (isNaN(imgId)) {
-      throw new Error('업로드된 이미지 ID가 유효하지 않습니다.');
-    }
-
-    return imgId;
+    return await mutation.mutateAsync(imageFile);
   };
 
   return {
     imageFile,
     imageUrl,
     handleImageChange,
-    uploadSelectedImage,
+    uploadSelectedImage, // async await 방식
+    ...mutation, // isPending, isError, isSuccess 등 그대로 꺼내쓸 수 있음
   };
 };
